+++ relay-server/src/testutils.rs
@@ -10,13 +10,13 @@ use relay_sampling::{DynamicSamplingContext, SamplingConfig};
 use relay_system::Addr;
 use relay_test::mock_service;
 
-use crate::actors::global_config::GlobalConfigHandle;
-use crate::actors::outcome::TrackOutcome;
-use crate::actors::processor::EnvelopeProcessorService;
-use crate::actors::project::ProjectState;
-use crate::actors::test_store::TestStore;
 use crate::envelope::{Envelope, Item, ItemType};
 use crate::extractors::RequestMeta;
+use crate::services::global_config::GlobalConfigHandle;
+use crate::services::outcome::TrackOutcome;
+use crate::services::processor::EnvelopeProcessorService;
+use crate::services::project::ProjectState;
+use crate::services::test_store::TestStore;
 
 pub fn state_with_rule_and_condition(
     sample_rate: Option<f64>,
@@ -115,11 +115,17 @@ pub fn create_test_processor(config: Config) -> EnvelopeProcessorService {
     #[cfg(feature = "processing")]
     let (_aggregator, _) = mock_service("aggregator", (), |&mut (), _| {});
 
+    #[cfg(feature = "processing")]
+    let redis = config
+        .redis()
+        .filter(|_| config.processing_enabled())
+        .map(|redis_config| relay_redis::RedisPool::new(redis_config).unwrap());
+
     EnvelopeProcessorService::new(
         Arc::new(config),
         GlobalConfigHandle::fixed(Default::default()),
         #[cfg(feature = "processing")]
-        None,
+        redis,
         outcome_aggregator,
         project_cache,
         upstream_relay,